+++ frontend/src/service/service-context.ts
@@ -1,6 +1,7 @@
 import ConfigService from "./config-service";
 import PlaylistService from "./playlist-service";
 import FileService from "./file-service";
+import AuthService from "./auth-service";
 
 export interface Services {
     config(): ConfigService;
@@ -8,6 +9,8 @@ export interface Services {
     playlist(): PlaylistService;
 
     file(): FileService;
+
+    auth(): AuthService;
 }
 
 class ServiceContextImpl implements Services {
@@ -15,6 +18,7 @@ class ServiceContextImpl implements Services {
     private readonly _configService: ConfigService = new ConfigService();
     private readonly _playlistService: PlaylistService = new PlaylistService();
     private readonly _fileService: FileService = new FileService();
+    private readonly _authService: AuthService = new AuthService();
 
     config() {
         return this._configService;
@@ -26,6 +30,10 @@ class ServiceContextImpl implements Services {
 
     file() {
         return this._fileService;
+    }
+
+    auth() {
+        return this._authService;
     }
 }