+++ src/auth/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod authenticator;
+pub(crate) mod password;
+pub(crate) mod user;